+++ ChangeLog.md
@@ -17,6 +17,19 @@ See docs/process.md for how version tagging works.
 
 Current Trunk
 -------------
+- `ALLOW_MEMORY_GROWTH` used to silently disable `ABORTING_MALLOC`. It now
+  just changes the default, which means you can pass `-s ABORTING_MALLOC=1` to
+  override the default, which was not possible before. (If you pass the flag
+  and don't want that behavior, stop passing the flag.) (#11131)
+- Change the factory function created by using the `MODULARIZE` build option to
+  return a Promise instead of the module instance. If you use `MODULARIZE` you
+  will need to wait on the returned Promise, using `await` or its `then`
+  callback, to get the module instance (#10697). This fixes some long-standing
+  bugs with that option which have been reported multiple times, but is a
+  breaking change - sorry about that. See detailed examples for the
+  current usage in `src/settings.js` on `MODULARIZE`.
+- A new `PRINTF_LONG_DOUBLE` option allows printf to print long doubles at full
+  float128 precision. (#11130)
 - `emscripten_async_queue_on_thread` has been renamed to
   `emscripten_dispatch_to_thread` which no longer implies that it is async -
   the operation is in fact only async if it is sent to another thread, while it
@@ -41,10 +54,6 @@ Current Trunk
 
 1.39.15: 05/06/2020
 -------------------
-- Change the factory function created by using the `MODULARIZE` build option to
-  return a Promise instead of the module instance. If you use `MODULARIZE` you
-  will need to wait on the returned Promise, using `await` or its `then`
-  callback, to get the module instance (#10697).
 - Add `--extern-pre-js` and `--extern-post-js` emcc flags. Files provided there
   are prepended/appended to the final JavaScript output, *after* all other
   work has been done, including optimization, optional `MODULARIZE`-ation,